Split pause point Unity IPC calls into their own file
pause_point_wait.go grew past the 500-line architecture limit
(TestProductionGoFilesStayFocused) once the expiry-extend call and
its bridge function were added. Move the raw Unity IPC callers
(queryPausePointStatusFromUnity, clearPausePointStatusFromUnity,
extendPausePointExpiryFromUnity, clearPausePointAfterWaitTimeout)
into pause_point_ipc.go, a cohesive group already sharing the same
probe-timeout pattern.
