+++ cli/project-runner/internal/projectrunner/pause_point_ipc.go
@@ -1,0 +1,88 @@
+package projectrunner
+
+import (
+	"context"
+	"encoding/json"
+
+	"github.com/hatayama/unity-cli-loop/common/clicontract"
+	"github.com/hatayama/unity-cli-loop/common/unityipc"
+)
+
+func queryPausePointStatusFromUnity(
+	ctx context.Context,
+	connection unityipc.Connection,
+	id string,
+) (pausePointStatusResponse, error) {
+	probeContext, cancel := context.WithTimeout(ctx, pausePointStatusProbeTimeout)
+	defer cancel()
+
+	result, err := unityipc.NewClient(connection, clicontract.ProjectRunnerVersion()).Send(
+		probeContext,
+		pausePointStatusCommandName,
+		map[string]any{"Id": id},
+	)
+	if err != nil {
+		return pausePointStatusResponse{}, err
+	}
+
+	response := pausePointStatusResponse{}
+	if err := json.Unmarshal(result, &response); err != nil {
+		return pausePointStatusResponse{}, err
+	}
+	return response, nil
+}
+
+func clearPausePointStatusFromUnity(
+	ctx context.Context,
+	connection unityipc.Connection,
+	id string,
+) (pausePointStatusResponse, error) {
+	probeContext, cancel := context.WithTimeout(ctx, pausePointStatusProbeTimeout)
+	defer cancel()
+
+	result, err := unityipc.NewClient(connection, clicontract.ProjectRunnerVersion()).Send(
+		probeContext,
+		pausePointClearStatusCommandName,
+		map[string]any{"Id": id},
+	)
+	if err != nil {
+		return pausePointStatusResponse{}, err
+	}
+
+	response := pausePointStatusResponse{}
+	if err := json.Unmarshal(result, &response); err != nil {
+		return pausePointStatusResponse{}, err
+	}
+	return response, nil
+}
+
+func extendPausePointExpiryFromUnity(
+	ctx context.Context,
+	connection unityipc.Connection,
+	id string,
+	minimumRemainingSeconds int,
+) (pausePointStatusResponse, error) {
+	probeContext, cancel := context.WithTimeout(ctx, pausePointStatusProbeTimeout)
+	defer cancel()
+
+	result, err := unityipc.NewClient(connection, clicontract.ProjectRunnerVersion()).Send(
+		probeContext,
+		pausePointExtendStatusCommandName,
+		map[string]any{"Id": id, "MinimumRemainingSeconds": minimumRemainingSeconds},
+	)
+	if err != nil {
+		return pausePointStatusResponse{}, err
+	}
+
+	response := pausePointStatusResponse{}
+	if err := json.Unmarshal(result, &response); err != nil {
+		return pausePointStatusResponse{}, err
+	}
+	return response, nil
+}
+
+func clearPausePointAfterWaitTimeout(ctx context.Context, connection unityipc.Connection, id string) {
+	clearContext, cancel := context.WithTimeout(ctx, pausePointStatusProbeTimeout)
+	defer cancel()
+	_, _ = clearPausePointStatus(clearContext, connection, id)
+}
