+++ cli/project-runner/internal/projectrunner/pause_point_wait.go
@@ -10,7 +10,6 @@ import (
 
 	clierrors "github.com/hatayama/unity-cli-loop/common/errors"
 
-	"github.com/hatayama/unity-cli-loop/common/clicontract"
 	"github.com/hatayama/unity-cli-loop/common/clicore"
 	"github.com/hatayama/unity-cli-loop/common/unityipc"
 )
@@ -450,83 +449,4 @@ func pausePointWaitStateForStatus(status string) pausePointWaitState {
 	default:
 		return ""
 	}
-}
-
-func queryPausePointStatusFromUnity(
-	ctx context.Context,
-	connection unityipc.Connection,
-	id string,
-) (pausePointStatusResponse, error) {
-	probeContext, cancel := context.WithTimeout(ctx, pausePointStatusProbeTimeout)
-	defer cancel()
-
-	result, err := unityipc.NewClient(connection, clicontract.ProjectRunnerVersion()).Send(
-		probeContext,
-		pausePointStatusCommandName,
-		map[string]any{"Id": id},
-	)
-	if err != nil {
-		return pausePointStatusResponse{}, err
-	}
-
-	response := pausePointStatusResponse{}
-	if err := json.Unmarshal(result, &response); err != nil {
-		return pausePointStatusResponse{}, err
-	}
-	return response, nil
-}
-
-func clearPausePointStatusFromUnity(
-	ctx context.Context,
-	connection unityipc.Connection,
-	id string,
-) (pausePointStatusResponse, error) {
-	probeContext, cancel := context.WithTimeout(ctx, pausePointStatusProbeTimeout)
-	defer cancel()
-
-	result, err := unityipc.NewClient(connection, clicontract.ProjectRunnerVersion()).Send(
-		probeContext,
-		pausePointClearStatusCommandName,
-		map[string]any{"Id": id},
-	)
-	if err != nil {
-		return pausePointStatusResponse{}, err
-	}
-
-	response := pausePointStatusResponse{}
-	if err := json.Unmarshal(result, &response); err != nil {
-		return pausePointStatusResponse{}, err
-	}
-	return response, nil
-}
-
-func extendPausePointExpiryFromUnity(
-	ctx context.Context,
-	connection unityipc.Connection,
-	id string,
-	minimumRemainingSeconds int,
-) (pausePointStatusResponse, error) {
-	probeContext, cancel := context.WithTimeout(ctx, pausePointStatusProbeTimeout)
-	defer cancel()
-
-	result, err := unityipc.NewClient(connection, clicontract.ProjectRunnerVersion()).Send(
-		probeContext,
-		pausePointExtendStatusCommandName,
-		map[string]any{"Id": id, "MinimumRemainingSeconds": minimumRemainingSeconds},
-	)
-	if err != nil {
-		return pausePointStatusResponse{}, err
-	}
-
-	response := pausePointStatusResponse{}
-	if err := json.Unmarshal(result, &response); err != nil {
-		return pausePointStatusResponse{}, err
-	}
-	return response, nil
-}
-
-func clearPausePointAfterWaitTimeout(ctx context.Context, connection unityipc.Connection, id string) {
-	clearContext, cancel := context.WithTimeout(ctx, pausePointStatusProbeTimeout)
-	defer cancel()
-	_, _ = clearPausePointStatus(clearContext, connection, id)
 }
